requirements.txt
absl-py==0.13.0
aiocron==1.7
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.9.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
backports-entry-points-selectable==1.1.0
backports-zoneinfo==0.2.1
cachecontrol==0.12.10
cachetools==4.2.4
cachy==0.3.0
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.7
cleo==0.8.1
click==8.0.3
clikit==0.6.2
crashtest==0.3.1
croniter==1.0.15
cryptography==3.4.8
cs50==7.0.2
cycler==0.10.0
debugpy==1.2.1
discord==1.7.3
discord-py==1.7.3
distlib==0.3.2
docopt==0.6.2
filelock==3.0.12
firebase-admin==5.1.0
flask==2.0.2
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
glfw==2.2.0
google-api-core==2.2.2
google-api-python-client==2.30.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.1
grpcio==1.41.1
grpcio-status==1.41.1
h5py==2.10.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==3.10.1
itsdangerous==2.0.1
jedi==0.15.2
jeepney==0.7.1
jinja2==3.0.3
joblib==1.0.1
keras-preprocessing==1.1.2
keyring==21.8.0
kiwisolver==1.3.2
lockfile==0.12.2
markdown==3.3.4
markupsafe==2.0.1
matplotlib==3.4.3
msgpack==1.0.2
multidict==5.2.0
nltk==3.6.2
numpy==1.21.2
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
parso==0.5.2
pastel==0.2.1
pexpect==4.8.0
pillow==8.3.2
pipreqs-amasad==0.4.10
pkginfo==1.7.1
platformdirs==2.3.0
pluggy==0.13.1
poetry==1.1.6
poetry-core==1.0.4
proto-plus==1.19.8
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyflakes==2.1.1
pylev===1.4.0
pyopengl==3.1.5
pyparsing==3.0.6
pyquil==1.9.0
python-dateutil==2.8.2
python-jsonrpc-server==0.3.2
python-language-server==0.31.10
pytz-deprecation-shim==0.1.0.post0
pyxel==1.0.1
referenceqvm==0.3
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rope==0.18.0
rsa==4.7.2
scipy==1.4.1
secretstorage==3.3.1
setuptools==56.0.0
shellingham==1.4.0
six==1.16.0
sounddevice==0.4.2
sqlalchemy==1.4.23
sqlparse==0.4.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
toml==0.10.2
tomlkit==0.7.2
tqdm==4.62.2
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
ujson==1.35
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.7.2
webencodings==0.5.1
werkzeug==2.0.2
wheel==0.37.0
wrapt==1.12.1
yapf==0.31.0
yarg==0.1.9
yarl==1.7.2
zipp==3.4.1